requirements.txt
ast
numpy==1.19.5
pubchempy
sklearn
pubchempy
matplotlib
rdkit-pypi
Pillow
IPython
pandas
opencv-python